Consolidate EmergencyStopException into osgar.exceptions
- Create a new `osgar/exceptions.py` file for shared OSGAR exceptions.
- Move `EmergencyStopException` from `osgar.followme` and various local
  definitions to the centralized `osgar.exceptions` module.
- Update core library, applications (SubT), and examples to use the
  centralized exception.
- Remove redundant duplicate definitions of `EmergencyStopException`.
