+++ osgar/ro2018.py
@@ -13,6 +13,7 @@ from osgar.lib.mathex import normalizeAnglePIPI
 from osgar.record import record
 
 from osgar.drivers.gps import INVALID_COORDINATES
+from osgar.exceptions import EmergencyStopException
 
 
 def geo_length(pos1, pos2):
@@ -31,9 +32,6 @@ def geo_angle(pos1, pos2):
 
 def latlon2xy(lat, lon):
     return int(round(lon*3600000)), int(round(lat*3600000))
-
-class EmergencyStopException(Exception):
-    pass
 
 class EmergencyStopMonitor:
     def __init__(self, robot):
