+++ osgar/test_ro2018.py
@@ -3,7 +3,8 @@ from unittest.mock import MagicMock
 import math
 
 from osgar.ro2018 import (geo_length, geo_angle,
-                          latlon2xy, EmergencyStopMonitor, EmergencyStopException)
+                          latlon2xy, EmergencyStopMonitor)
+from osgar.exceptions import EmergencyStopException
 
 
 class RO2018Test(unittest.TestCase):
